backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
sqlalchemy==2.0.36
python-multipart==0.0.19
python-dotenv==1.0.1
google-generativeai==0.8.4
pandas==2.2.3
scikit-learn==1.6.1
xgboost==2.1.3
joblib==1.4.2
numpy==1.26.4
aiofiles==24.1.0
Pillow==11.1.0
PyPDF2==3.0.1
